Fixed issue with MILP method
Previously, MILP could force reactions active which had negative ub, therefore changing preexisting bounds in an undesirable way. Updated the method so that now, if the reaction can't be forced forward (for example) it will be forced to be not reversed. This behavior is still not well documented. TODO: Document this more clearly.

Since the model can fairly easily handle infeasiblity (by ignoring the restrictions that could theoretically come from the infeasible solutions), it is okay for the MILPs to cause infeasibility.

There may specifically be issues with maintanance reactions bounds being turned off which could be a problem (TODO: Investigate this).