Added tests for imat methods
Added tests for iMAT, and updated iMAT code to correct any errors.
Additionally, corrected error in parse_gpr, changing the || regex to use
a raw string to avoid error with invalid escape sequences.

There seems to be an error in either COBRApy or optlang, that copying
models (i.e. model.copy) will change the type of variables, from "binary" to "continous". This is handled in this package by forcing known binary variables to be binary, but it is an awkard solution. Should find a minimal reproducible example and submit a github issue.

Future TODO:
- Current tests for add constraints and objective just check that these
  exist, not their correctness. Should add test cases to adress this
  issue.
- Write tests for the model creation methods based on the iMAT
  functionality.