Added tests against GLPK and Gurobi solvers
Added tests against additional solvers, specifically GLPK and GUROBI.
Due to a depracation warning coming from gurobi (via optlang) there are
a lot of warnings in the tests. Default solver was set to GLPK to stop
these warnings for unrelated tests, so the only tests that use gurobi
are the ones directly testing that interface.